lab5/pyproject.toml
[project]
name = "lab5"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.6",
    "django-unfold>=0.65.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
]
